+++ m1_csrc/m1_t1000_version.h
@@ -7,7 +7,7 @@
 
 #define T1000_VERSION_MAJOR        0
 #define T1000_VERSION_MINOR        2
-#define T1000_VERSION_PATCH        0
+#define T1000_VERSION_PATCH        1
 
 #define T1000_STR_HELPER(x)        #x
 #define T1000_STR(x)               T1000_STR_HELPER(x)